+++ src/main/java/ar/edu/itba/ss/integrators/Integrator.java
@@ -1,10 +1,11 @@
 package ar.edu.itba.ss.integrators;
 
 import ar.edu.itba.ss.Oscillator;
+import java.math.BigDecimal;
 
 public interface Integrator {
 
-    void initialize(Oscillator osc, double x, double v, double dt);
-    double[] step(double x, double v, double t);
+    void initialize(Oscillator osc, BigDecimal x, BigDecimal v, BigDecimal dt);
+    BigDecimal[] step(BigDecimal x, BigDecimal v, BigDecimal t);
 
 }
